requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-admin-sortable==2.0
django-braces==1.8.1
django-cors-headers==1.1.0
django-extensions==1.5.6
django-oauth-toolkit==0.9.0
django-toolbelt==0.0.1
djangorestframework==3.2.3
gunicorn==19.3.0
inflection==0.3.1
mysqlclient==1.3.6
oauthlib==1.0.3
psycopg2==2.6.1
PyJWT==1.4.0
python-social-auth==0.2.12
python3-openid==3.0.6
PyYAML==3.11
requests==2.7.0
requests-oauthlib==0.5.0
rest-framework-ember==1.3.2
six==1.9.0
static3==0.6.1
wheel==0.24.0
